[add]select user by name

+++ src/main/java/gg/jominsubyungsin/domain/entity/UserEntity.java
@@ -23,7 +23,7 @@ public class UserEntity {
   private String email;
 
   @Column(nullable = false)
-  private int mailAccess;
+  private byte mailAccess;
 
   @JsonIgnore
   @Column(nullable = false)
